[BUD-65] Add utils/apollo-client
